LayoutTests/fast/mediastream/enabled-expected.txt

Tests that navigator.webkitGetUserMedia is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof navigator.mediaDevices == 'object' is true
PASS typeof navigator.mediaDevices.getUserMedia == 'function' is true
PASS hasGetUserMediaProperty() is true
PASS 'mediaDevices' in navigator is true
PASS 'getUserMedia' in navigator.mediaDevices is true
PASS navigator.hasOwnProperty('mediaDevices') is false
PASS successfullyParsed is true

TEST COMPLETE
